requirements.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
bleach==4.1.0
build==0.7.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dirhunt==0.8.1
docutils==0.18.1
google==3.0.0
greenlet==1.1.2
humanize==4.0.0
idna==3.3
importlib-metadata==4.11.2
keyring==23.5.0
packaging==21.3
pep517==0.12.0
pkginfo==1.8.2
proxy-db==0.3.1
pycodestyle==2.8.0
Pygments==2.11.2
pyparsing==3.0.7
python-dotenv==0.19.2
readme-renderer==34.0
requests==2.27.1
requests-mock==1.9.3
requests-toolbelt==0.9.1
rfc3986==2.0.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
toml==0.10.2
tomli==2.0.1
tqdm==4.63.0
twine==3.8.0
urllib3==1.26.8
webencodings==0.5.1
zipp==3.7.0
